+++ src/app/(service)/myhistory/hooks/useHistoryBoardData.ts
@@ -5,7 +5,11 @@ import { useMemo } from 'react';
 import { useQueries } from '@tanstack/react-query';
 
 import { taskQueryOptions, teamQueryOptions } from '@/api/queryOptions';
-import type { MyHistoryCompletedTaskRecord } from '@/app/(service)/myhistory/types';
+import type {
+  HistoryTaskListDetailSource,
+  HistoryTeamDetail,
+  UseHistoryBoardDataParams,
+} from '@/app/(service)/myhistory/types';
 import {
   getCompletedDateKeys,
   getHistoryTaskListDescriptors,
@@ -13,8 +17,6 @@ import {
 import {
   getHistorySections,
   getHistorySummaryData,
-  type HistoryTaskListDetailSource,
-  type HistoryTeamDetail,
   toHistoryCurrentUserId,
   toHistoryTaskListDetailSource,
   toHistoryTeamDetail,
@@ -22,10 +24,11 @@ import {
 } from '@/app/(service)/myhistory/utils/myHistoryData';
 import { useMeQuery, useMyMembershipsQuery } from '@/hooks/useUser';
 
-type UseHistoryBoardDataParams = {
-  activeFilterId: string | null;
-  completedTasks: readonly MyHistoryCompletedTaskRecord[];
-};
+function getTaskListDescriptorKey(
+  descriptor: ReturnType<typeof getHistoryTaskListDescriptors>[number],
+) {
+  return `${descriptor.teamId}-${descriptor.taskListId}-${descriptor.dateKey}`;
+}
 
 export default function useHistoryBoardData({
   activeFilterId,
@@ -36,23 +39,37 @@ export default function useHistoryBoardData({
     isLoading: isMeLoading,
     isError: isMeError,
   } = useMeQuery();
+
   const {
     data: membershipsData,
     isLoading: isMembershipsLoading,
     isError: isMembershipsError,
   } = useMyMembershipsQuery();
+
   const currentUserId = useMemo(() => toHistoryCurrentUserId(meData), [meData]);
+
   const teams = useMemo(
     () => toHistoryTeams(membershipsData),
     [membershipsData],
   );
+
   const completedDateKeys = useMemo(
     () => getCompletedDateKeys(completedTasks),
     [completedTasks],
   );
 
+  const uniqueTeams = useMemo(() => {
+    const map = new Map<string | number, (typeof teams)[number]>();
+
+    teams.forEach((team) => {
+      map.set(team.id, team);
+    });
+
+    return [...map.values()];
+  }, [teams]);
+
   const teamDetailQueries = useQueries({
-    queries: teams.map((team) => teamQueryOptions.detail(team.id)),
+    queries: uniqueTeams.map((team) => teamQueryOptions.detail(team.id)),
   });
 
   const teamDetails = useMemo(
@@ -70,14 +87,27 @@ export default function useHistoryBoardData({
     [completedDateKeys, teamDetails],
   );
 
+  const uniqueTaskListDescriptors = useMemo(() => {
+    const map = new Map<string, (typeof taskListDescriptors)[number]>();
+
+    taskListDescriptors.forEach((descriptor) => {
+      map.set(getTaskListDescriptorKey(descriptor), descriptor);
+    });
+
+    return [...map.values()];
+  }, [taskListDescriptors]);
+
   const taskListDetailQueries = useQueries({
-    queries: taskListDescriptors.map((descriptor) =>
+    queries: uniqueTaskListDescriptors.map((descriptor) =>
       taskQueryOptions.taskListDetail(
         descriptor.teamId,
         descriptor.taskListId,
         { date: descriptor.dateKey },
         {
-          enabled: taskListDescriptors.length > 0,
+          enabled:
+            Boolean(descriptor.teamId) &&
+            Boolean(descriptor.taskListId) &&
+            Boolean(descriptor.dateKey),
         },
       ),
     ),
@@ -85,7 +115,7 @@ export default function useHistoryBoardData({
 
   const taskListSources = useMemo(
     () =>
-      taskListDescriptors.reduce<HistoryTaskListDetailSource[]>(
+      uniqueTaskListDescriptors.reduce<HistoryTaskListDetailSource[]>(
         (sources, descriptor, index) => {
           const data = taskListDetailQueries[index]?.data;
 
@@ -99,13 +129,14 @@ export default function useHistoryBoardData({
         },
         [],
       ),
-    [taskListDescriptors, taskListDetailQueries],
+    [uniqueTaskListDescriptors, taskListDetailQueries],
   );
 
   const historySections = useMemo(
     () => getHistorySections(completedTasks, taskListSources, activeFilterId),
     [activeFilterId, completedTasks, taskListSources],
   );
+
   const summaryData = useMemo(
     () => getHistorySummaryData(currentUserId, teamDetails, taskListSources),
     [currentUserId, taskListSources, teamDetails],
