--- src/app/(service)/myhistory/types/date.ts
@@ -1,15 +1,0 @@
-export type MyHistoryDateSelectionMode = 'all' | 'month' | 'range';
-
-export type MyHistoryResolvedDateRange = {
-  endDate: Date;
-  startDate: Date;
-};
-
-export type MyHistoryDateRange = MyHistoryResolvedDateRange & {
-  mode: MyHistoryDateSelectionMode;
-};
-
-export type MyHistoryDraftDateRange = {
-  endDate: Date | null;
-  startDate: Date | null;
-};
