+++ src/components/layout/header/components/Header.tsx
@@ -4,6 +4,8 @@
 
 'use client';
 
+import { useSyncExternalStore } from 'react';
+
 import Image from 'next/image';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
@@ -15,17 +17,40 @@ import useMobileSidebar from '@/components/layout/header/hooks/useMobileSidebar'
 import useLayoutAuthState from '@/components/layout/hooks/useLayoutAuthState';
 import { ROUTES } from '@/constants/ROUTES';
 
+function subscribeMounted(callback: () => void) {
+  callback();
+
+  return () => {};
+}
+
+function getClientSnapshot() {
+  return true;
+}
+
+function getServerSnapshot() {
+  return false;
+}
+
 export default function Header() {
   const pathname = usePathname();
   const layoutAuthState = useLayoutAuthState(pathname);
+
   const { handleClose, handleToggle, isRendered, isVisible, menuButtonRef } =
     useMobileSidebar();
+
+  const isMounted = useSyncExternalStore(
+    subscribeMounted,
+    getClientSnapshot,
+    getServerSnapshot,
+  );
+
+  const canShowAuthUi = isMounted && layoutAuthState.isAuthenticated;
 
   return (
     <>
       <header className="md:hidden sticky top-0 z-40 flex h-13 w-full items-center border-b border-background-tertiary bg-background-inverse px-4">
         <div className="flex items-center gap-3">
-          {layoutAuthState.isAuthenticated && (
+          {canShowAuthUi && (
             <button
               ref={menuButtonRef}
               type="button"
@@ -56,7 +81,7 @@ export default function Header() {
           </Link>
         </div>
 
-        {layoutAuthState.isAuthenticated ? (
+        {canShowAuthUi ? (
           <ProfileMenuDropdown
             className="ml-auto"
             trigger={
@@ -93,7 +118,7 @@ export default function Header() {
         )}
       </header>
 
-      {layoutAuthState.isAuthenticated && (
+      {canShowAuthUi && (
         <MobileSidebarDrawer
           isRendered={isRendered}
           isVisible={isVisible}
